chore: move auth guard to middleware
